backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
pydantic==2.11.10
anyio==4.11.0
starlette==0.48.0
openai==1.52.2
websockets==12.0
python-dotenv==1.1.1
httpx==0.27.2

